RA-902 Automation | Add Social Pin Phone Contact
- Refactor Playwright tests and update configurations for improved application group handling. - Add application group tests for social pins
