+++ integration_tests/playwright.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   testDir: './playwright_tests/e2e',
   fullyParallel: false,
   forbidOnly: !!process.env.CI,
-  retries: process.env.CI ? 1 : 1,
+  retries: process.env.CI ? 1 : 0,
   workers: 1,
   reporter: [['html', { outputFolder: './playwright_tests/reports' }], ['list']],
   use: {
