+++ integration_tests/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
-    "target": "es5",
+    "target": "ES2020",
+    "moduleResolution": "node",
     "noEmit": true,
-    "lib": ["es5", "dom", "es2015.promise"],
+    "lib": ["ES2020", "dom"],
     "types": ["cypress", "express", "express-session"],
     "esModuleInterop": true,
     "resolveJsonModule": true,
